+++ backend/src/zod-validation/zod-schemas/formQuestionSchema.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod';
-import { QuestionType } from '../types/questionType';
+import { QuestionType } from '../../types/questionType';
 
 // Full Schema for FormQuestion (POST requests)
 export const formQuestionSchema = z.object({
